+++ Source/ACE.Server/Network/GameAction/Actions/GameActionAddChannel.cs
@@ -1,4 +1,4 @@
-﻿using ACE.Entity.Enum;
+using ACE.Entity.Enum;
 
 namespace ACE.Server.Network.GameAction.Actions
 {
@@ -7,7 +7,7 @@ namespace ACE.Server.Network.GameAction.Actions
         [GameAction(GameActionType.AddChannel)]
         public static void Handle(ClientMessage message, Session session)
         {
-            var chatChannelID = (GroupChatType)message.Payload.ReadUInt32();
+            var chatChannelID = (Channel)message.Payload.ReadUInt32();
             // Probably need some IsAdvocate and IsSentinel type thing going on here as well. leaving for now
             if (!session.Player.IsAdmin && !session.Player.IsArch && !session.Player.IsPsr)
                 return;